added upgrade logic (not finished, 1/2 step done)

+++ Drill Game/Assets/Scripts/Blocks/Item.cs
@@ -5,11 +5,11 @@ namespace Blocks
 {
     public class Item : MonoBehaviour
     {
-        [SerializeField] private int _cost = 1;
+        [SerializeField] private float _cost = 1;
         private const float PeakAmplitude = 5;
 
         public bool IsCollected { get; private set; }
-        public int Cost => _cost;
+        public float Cost => _cost;
         
         public void Collect(Transform parent, Vector3 localTargetPosition, float duration = 0.3f)
         {
